Convert angular component to react container.
