+++ galaxyui/src/app/authors-react/react-components/patternfly-pager.tsx
@@ -9,42 +9,25 @@ interface IProps {
 }
 
 interface IState {
-    page?: number;
     perPage?: number;
 }
 
-class PaginationConfig {
+export class PaginationConfig {
     pageNumber?: number;
     totalItems?: number;
     // pageSizeIncrements?: Array<number> = [5, 10, 20, 40, 80, 100];
     pageSize?: number;
 }
 
-export class PagerPF extends React.Component<IProps, IState> {
+export class PagerPF extends React.Component<IProps, {}> {
     pageSizeIncrements = [5, 10, 20, 40, 80, 100];
 
-    constructor(props) {
-        super(props);
-        this.state = {
-            page: this.props.config.pageNumber,
-            perPage: this.props.config.pageSize,
-        };
-
-        // if (this.props.config.pageSizeIncrements === undefined) {
-        //     this.props.config.pageSizeIncrements = [5, 10, 20, 40, 80, 100];
-        // }
-    }
-
     onPageSet(count) {
-        this.setState({ page: count }, () =>
-            this.props.onPageNumberChange({ pageNumber: this.state.page }),
-        );
+        this.props.onPageNumberChange({ pageNumber: count });
     }
 
     onPerPageSelect(count) {
-        this.setState({ perPage: count }, () =>
-            this.props.onPageSizeChange({ pageSize: this.state.perPage }),
-        );
+        this.props.onPageSizeChange({ pageSize: count });
     }
 
     render() {
@@ -52,8 +35,8 @@ export class PagerPF extends React.Component<IProps, IState> {
             <Paginator
                 viewType='list'
                 pagination={{
-                    page: this.state.page,
-                    perPage: this.state.perPage,
+                    page: this.props.config.pageNumber,
+                    perPage: this.props.config.pageSize,
                     perPageOptions: this.pageSizeIncrements,
                 }}
                 itemCount={this.props.config.totalItems}
